Pixel Navbar: Hide OPA when not pressed
Extracted pieces of code from the below code and refactor it into my previous changes.
It should hide the 4 dots when not pressed, but it doesn't mess with the halo.

Original commit:
fwb: Pixel Navbar: Hide opa when not pressed, also remove Halo

Since we use the AOSP Navbar Icon, that is transparent in the middle,
we don't want the opa to always show. Instead, only show them when a
press is getting executed (when they need to be shown for the animation).

fadeInButton and fadeOutButton are based on this Stackoverflow answer:
http://stackoverflow.com/a/22669115

Change-Id: Iaee1c31aaa157d5af55ad45b263be2dd37aafc22
Signed-off-by: mydongistiny <jaysonedson@gmail.com>